# content/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunIdRequest": {
        "additionalProperties": false,
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "RunIdRequest",
        "type": "object"
      },
...
      "Stage1AdvanceRequest": {
        "additionalProperties": false,
        "properties": {
          "absolute_index": {
            "minimum": 1.0,
            "title": "Absolute Index",
            "type": "integer"
          },
          "player_intents": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Player Intents",
            "type": "object"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id",
          "absolute_index"
        ],
        "title": "Stage1AdvanceRequest",
        "type": "object"
      },
      "Stage1AdvanceResponse": {
        "additionalProperties": false,
        "properties": {
          "absolute_index": {
            "title": "Absolute Index",
            "type": "integer"
          },
          "duplicate": {
            "default": false,
            "title": "Duplicate",
            "type": "boolean"
          },
          "quiet": {
            "default": false,
            "title": "Quiet",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/Stage1SceneOutcome"
            },
            "title": "Scenes",
            "type": "array"
          },
          "snapshot_id": {
            "format": "uuid",
            "title": "Snapshot Id",
            "type": "string"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "world_id",
          "absolute_index",
          "snapshot_id"
        ],
        "title": "Stage1AdvanceResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "worldsim",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/stage1/advance": {
      "post": {
        "operationId": "advance_api_v1_stage1_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Stage1AdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stage1AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/characters": {
      "get": {
        "operationId": "list_characters_api_v1_stage1_characters_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterSummary"
                  },
                  "title": "Response List Characters Api V1 Stage1 Characters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Characters",
        "tags": [
          "stage1"
        ]
      },
      "post": {
        "description": "Create a simulation character with its identity card.",
        "operationId": "create_character_view_api_v1_stage1_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character View",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/characters/{character_id}": {
      "get": {
        "operationId": "get_character_api_v1_stage1_characters__character_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/model-runs": {
      "get": {
        "operationId": "list_model_runs_api_v1_stage1_model_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "phase_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Phase Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelRunView"
                  },
                  "title": "Response List Model Runs Api V1 Stage1 Model Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Model Runs",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/party": {
      "get": {
        "description": "Party sheets are shared party knowledge: full roster for both roles.",
        "operationId": "party_roster_api_v1_stage1_party_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyRosterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Party Roster",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/party/begin": {
      "post": {
        "description": "Seat the player's adventurer (explicit stats or the auto build).",
        "operationId": "begin_party_member_api_v1_stage1_party_begin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartyBeginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyMemberView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Begin Party Member",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/party/{member_id}/link": {
      "post": {
        "description": "Bind an existing roster row to a real character, version-checked.",
        "operationId": "link_party_member_api_v1_stage1_party__member_id__link_post",
        "parameters": [
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Member Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartyLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyMemberView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Party Member",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/pause": {
      "post": {
        "operationId": "pause_api_v1_stage1_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Pause Api V1 Stage1 Pause Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/resume": {
      "post": {
        "operationId": "resume_api_v1_stage1_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Resume Api V1 Stage1 Resume Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/scenes": {
      "get": {
        "operationId": "list_scenes_api_v1_stage1_scenes_get",
        "parameters": [
          {
            "in": "query",
            "name": "phase_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Phase Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SceneSummary"
                  },
                  "title": "Response List Scenes Api V1 Stage1 Scenes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scenes",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/scenes/{scene_id}": {
      "get": {
        "operationId": "get_scene_api_v1_stage1_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scene",
        "tags": [
          "stage1"
        ]
      }
    },
    "/api/v1/stage1/scenes/{scene_id}/narration": {
      "get": {
        "operationId": "get_narration_api_v1_stage1_scenes__scene_id__narration_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BeatView"
                  },
                  "title": "Response Get Narration Api V1 Stage1 Scenes  Scene Id  Narration Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Narration",
        "tags": [
          "stage1"
        ]
      }
    },
...
  }
}